Restructure framework modules
