+++ src/node/worker.py
@@ -1,7 +1,8 @@
-from ..node import Node
+from . import Node
+from .master import Master
 
-from .rx_pipe import get_supported_rx_pipe
-from .task_executor import TaskExecutor
+from ..rx_pipe import get_supported_rx_pipe
+from ..task_executor import TaskExecutor
 
 class Worker(Node):
     def __init__(self, task_executor : TaskExecutor):
@@ -13,7 +14,7 @@ class Worker(Node):
                                              self.config_manager.listen_port,
                                              legacy_asyncio = self.check_asyncio_use_legacy())
         self.rx_pipe.set_logger(self.logger)
-        if self.task_deployer is not None:
+        if isinstance(self, Master):
             self.rx_pipe.task_deployer = self.task_deployer
 
     def start(self):
