+++ src/task_deployer/task_deployer_latest.py
@@ -1,6 +1,6 @@
 import asyncio
 
-from .master import TaskDeployer
+from ..task_deployer import TaskDeployer
 
 class TaskDeployerLatest(TaskDeployer):
     """Task deployer implementation using the current asyncio interface.
